+++ web-vue/src/api/build-info.js
@@ -259,7 +259,7 @@ export const releaseMethodMap = {
   2: "项目",
   3: "SSH",
   4: "本地命令",
-  5: "Docker",
+  5: "Docker镜像",
 };
 
 export const triggerBuildTypeMap = {